refactor pastas
